Fixed the flicker of the display
The FPGA in the 1014D expects 0x78 written to command 0x38 to set the backlight PWM to fully on. Since the FPGA is not clocked yet due to the MS5351M clock synthesizer not being initialized yet, it can't generate a proper PWM signal.

Changed the FPGA backlight write to only send a byte value.

+++ Test code/fnirsi_1014d_firmware_backup/fwb_fpga_control.h
@@ -52,7 +52,7 @@ uint16 fpga_read_short(void);
 
 void fpga_write_int(uint32 data);
 
-void fpga_set_backlight_brightness(uint16 brightness);
+void fpga_set_backlight_brightness(uint8 brightness);
 
 uint16 fpga_get_version(void);
